requirements.txt
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
audioread==3.0.0
cffi==1.15.1
charset-normalizer==3.2.0
ci-sdr==0.0.2
click==8.1.7
ConfigArgParse==1.7
ctc-segmentation==1.7.4
Cython==3.0.0
decorator==5.1.1
Distance==0.1.3
editdistance==0.6.2
einops==0.6.1
#espnet==202308
git+https://github.com/airenas/espnet.git@v.202511-ai#egg=espnet
espnet-tts-frontend==0.0.3
fast-bss-eval==0.1.3
filelock==3.12.2
g2p-en==2.1.0
h5py==3.8.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
inflect==6.0.5
jaconv==0.3.4
jamo==0.4.1
joblib==1.3.2
kaldiio==2.18.0
librosa==0.9.2
llvmlite==0.39.1
nltk==3.8.1
numba==0.56.4
numpy==1.21.6
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.1
platformdirs==3.10.0
pooch==1.7.0
protobuf==3.20.1
pycparser==2.21
pydantic==1.10.12
pypinyin==0.44.0
pytorch-wpe==0.0.1
pyworld==0.3.4
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
resampy==0.4.2
scikit-learn==1.0.2
scipy==1.7.3
#sentencepiece==0.1.97
soundfile==0.12.1
threadpoolctl==3.1.0
torch-complex==0.4.3
tqdm==4.66.1
typeguard==2.13.3
typing_extensions==4.7.1
Unidecode==1.3.6
urllib3==2.0.4
zipp==3.15.0
uvicorn==0.22.0
Werkzeug==1.0.1
fastapi==0.101.1
starlette_exporter==0.16.0
espnet-model-zoo==0.1.7
git+https://github.com/airenas/smart_load_balancer.git@v0.2.37#egg=smart_load_balancer
 msgpack==1.1.2
